Convert find-messages.py into a gcc plugin that doesn't use python.
gcc-plugin-python appears to have been abandoned for some time now.  I
haven't been able to get it to build for a year or two, and the current
build failures are outside my ability to diagnose.

But, we don't use a lot of the features of that project in the first
place.  So instead, convert find-messages.py into a gcc plugin.  There's
probably a lot of work to go - the second step isn't implemented, more
code could probably be moved into separate files, memory management is
probably lacking, etc.